Added rb to modules, refactored with pylint/pep8, changed from tabs to spaces
