+++ src/rbai/rbai.py
@@ -8,16 +8,16 @@ from rbbase.rbbase import RBBaseGameObject, RB2DPosition
 
 class RBAI(RBBaseGameObject):
 
-	def __init__(self, x, y):
-		self._pos = RB2DPosition(x, y)
-		self._behaviour = None
+    def __init__(self, x, y):
+        self._pos = RB2DPosition(x, y)
+        self._behaviour = None
 
-	def update(self):
-		if self._behaviour:
-			self._behaviour.updatePosition(self._pos)
+    def update(self):
+        if self._behaviour:
+            self._behaviour.updatePosition(self._pos)
 
-	def draw(self, graphics):
-		pass
+    def draw(self, graphics):
+        pass
 
-	def setBehaviour(self, behaviour):
-		self._behaviour = behaviour
+    def setBehaviour(self, behaviour):
+        self._behaviour = behaviour
