+++ src/rbbase/rbbase.py
@@ -1,24 +1,24 @@
 class RBBaseGameObject(object):
 
-	def __init__(self):
-		pass
+    def __init__(self):
+        pass
 
 
 class RB2DPosition(object):
 
-	def __init__(self, x, y):
-		self.setPos(x, y)
+    def __init__(self, x, y):
+        self.setPos(x, y)
 
-	def setPos(self, x, y):
-		self._x = x
-		self._y = y
+    def setPos(self, x, y):
+        self._x = x
+        self._y = y
 
-	def getX(self):
-		return self._x
+    def getX(self):
+        return self._x
 
-	def getY(self):
-		return self._y
+    def getY(self):
+        return self._y
 
-	def movePos(self, x, y):
-		self._x += x
-		self._y += y
+    def movePos(self, x, y):
+        self._x += x
+        self._y += y
