--- src/rbphysics/collision.py
@@ -1,5 +1,0 @@
-class Collision(object):
-
-	def collideRectangleToRectangle(rectangleOne, rectangleTwo):
-		return not(rectangleOne._left > rectangleTwo._right or rectangleOne._right < rectangleTwo._left
-		or rectangleOne._top > rectangleTwo._bottom or rectangleOne._bottom < rectangleTwo._top)
